Renamed aws-security-credentials.json to aws-security-credentials-fake.json
