+++ test/fixtures/aws-security-credentials-fake.json
