+++ test/test.awsrequestsigner.ts
@@ -36,7 +36,7 @@ describe('AwsRequestSigner', () => {
   let clock: sinon.SinonFakeTimers;
   // Load AWS credentials from a sample security_credentials response.
   // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const awsSecurityCredentials = require('../../test/fixtures/aws-security-credentials.json');
+  const awsSecurityCredentials = require('../../test/fixtures/aws-security-credentials-fake.json');
   const accessKeyId = awsSecurityCredentials.AccessKeyId;
   const secretAccessKey = awsSecurityCredentials.SecretAccessKey;
   const token = awsSecurityCredentials.Token;
